Moving auth folder to features
