+++ client/src/app/features/auth/pages/login-form-page/login-form-page.component.ts
@@ -7,14 +7,14 @@ import {
 } from '@angular/forms';
 import { Router, RouterLink } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
-import { AccountService } from '../../../core/services/account.service';
-import { TextInputComponent } from '../../../shared/components/forms/text-input/text-input.component';
+import { AccountService } from '../../../../core/services/account.service';
+import { TextInputComponent } from '../../../../shared/components/forms/text-input/text-input.component';
 
 @Component({
   selector: 'app-login-form',
   imports: [ReactiveFormsModule, TextInputComponent, RouterLink],
-  templateUrl: './login-form.component.html',
-  styleUrls: ['./login-form.component.css'],
+  templateUrl: './login-form-page.component.html',
+  styleUrls: ['./login-form-page.component.css'],
 })
 export class LoginFormComponent {
   accountService = inject(AccountService);
